+++ tests/integration/components/t-add-test.js
@@ -10,15 +10,17 @@ module('Integration | Component | t add', function(hooks) {
   test('it renders', async function(assert) {
     assert.expect(2);
 
-    await render(hbs`{{t-add}}`);
+    await render(hbs`
+      <TAdd />
+    `);
 
     assert.dom('button').hasText('add item');
 
     // Template block usage:
     await render(hbs`
-      {{#t-add}}
+      <TAdd>
         template block text
-      {{/t-add}}
+      </TAdd>
     `);
 
     assert.dom('button').hasText('add item');
@@ -27,7 +29,9 @@ module('Integration | Component | t add', function(hooks) {
   test('it creates an add transformicon with defaults', async function(assert) {
     assert.expect(6);
 
-    await render(hbs`{{t-add}}`);
+    await render(hbs`
+      <TAdd />
+    `);
     percySnapshot(assert);
 
     assert.dom('button').hasAttribute('type', 'button');
@@ -41,7 +45,9 @@ module('Integration | Component | t add', function(hooks) {
   test('it creates an add transformicon with `is-added=true`', async function(assert) {
     assert.expect(1);
 
-    await render(hbs`{{t-add is-added=true}}`);
+    await render(hbs`
+      <TAdd @is-added={{true}} />
+    `);
     percySnapshot(assert);
 
     assert.dom('button').hasClass('tcon-transform');
@@ -50,7 +56,9 @@ module('Integration | Component | t add', function(hooks) {
   test('it creates an add transformicon with a non-default animation `a="check"`', async function(assert) {
     assert.expect(1);
 
-    await render(hbs`{{t-add a="check"}}`);
+    await render(hbs`
+      <TAdd @a="check" />
+    `);
 
     assert.dom('button').hasClass('tcon-plus--check');
   });
@@ -58,7 +66,11 @@ module('Integration | Component | t add', function(hooks) {
   test('user can click on the transformicon', async function(assert) {
     assert.expect(2);
 
-    await render(hbs`{{t-add id="t-add"}}`);
+    // FIXME: https://github.com/rwjblue/ember-angle-bracket-invocation-polyfill/issues/4#issue-328822657
+    // Currently necessary to use `<TAdd @id="t-add" />` syntax when specifying an `id` but the `aid` prefix shouldn't be necessary -- should be able to say `<TAdd @id="t-add" />`. Fix after this issue is closed.
+    await render(hbs`
+      <TAdd @id="t-add" />
+    `);
 
     assert.dom('#t-add').hasNoClass('tcon-transform');